backend/requirements.txt
fastapi
uvicorn[standard]
langchain-core
selenium
webdriver-manager
pydantic
langgraph
langchain-openai
python-dotenv
